Convert http_backend to an interface
By converting the http_backend into a implementation of a common
interface, it allows flexibility of using other HTTP clients or
remote browsers to drive HTTP requests. This commit adds the HTTPDriver
interface which provides the same functionality as the http_backend and
updates the Collector to communicate the the http_backend through the
interface methods.