Defines an HTTPBackend interface to use with Colly.
This interface is designed to support using different backends for fetching data, specifically to support using chromedp as a backend to support parsing sites that rely on .js to populate data.

It's based off of work by @mazhigali in https://github.com/gocolly/colly/pull/503/files, with some simplifications and tweaks to naming conventions.

The intent is to provide a fast follow with an implementation of a chromedp based backend as well as modifications to colly.go to support setting up with different backends, with this first PR setting up for that with a smaller changeset that's hopefully easier to review.