Remove tabs and newlines from URLs
This might sound weird, but both URL standard[1] specifies it,
and browsers do that as well.

Although the standard specifies it as a "validation error",
this is not a hard error.

This actually happens in the wild: as of now, this Google's page[2]
has the following fragment:

    <a class="glue-header__link"
                                  href="/intl/ru_ALL
    /drive/download/"
    >

Yes, the newline here is in the middle of the link, and browsers
do ignore it.

[1] https://url.spec.whatwg.org/#concept-basic-url-parser
[2] https://www.google.com/intl/ru/drive/download/