Use http.NewRequest instead of &http.Request{...}
http.NewRequest already does exactly what setRequestBody did.
In fact, scrape and setRequestBody used to contain parts of
http.NewRequest implementation.

119df7afac implies that NewRequest is slower than assembling
http.Request object manually, likely because it avoids URL
reparsing. Indeed, this change brings some slowdown:

    name           old time/op  new time/op  delta
    OnHTML-24       129µs ± 0%   131µs ± 1%  +1.69%  (p=0.029 n=4+4)
    OnXML-24        130µs ± 0%   132µs ± 1%  +1.69%  (p=0.029 n=4+4)
    OnResponse-24  93.4µs ± 0%  95.5µs ± 1%  +2.30%  (p=0.029 n=4+4)

However, the benchmarks are somewhat unrealistic as the HTML page is
extremely simple, the web server runs on localhost, and the data is
discarded (rather than e.g. stored somewhere). In practice,
impact of this is going to be neglible, and code complexity and
code duplication is not really worth it.