Use request.GetBody for request hashing
streamToByte has an ugly type switch that tries to rewind the body so it
could be read again.

However, http.NewRequest already does something similar: when Body type
allows, GetBody is populated so request can be retried. So just use it
instead of trying to reimplement the body rewinding, and remove
streamToByte function altogether.