Refactor existing caching mechanism into pluggable system
Create an interface `Cache`, similiar to `storage.Storage` that
will allow for pluggable caching extensions. Caching subsystem is
created with full access to Request and Response objects, to potentialy
make more complex caching decisions based on both request and response
headers. Because of the need to access those objects, `Cache` is not
created in its own package like `storage`, but in root `colly`.

Closes #103